% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{BioDIGS_DNA_conc_data}
\alias{BioDIGS_DNA_conc_data}
\title{Produce DNA concentration data in a clean format for use in R.}
\usage{
BioDIGS_DNA_conc_data(info = TRUE)
}
\arguments{
\item{info}{if set to TRUE, print important information to the console. This can be disabled with \code{info = FALSE}.}
}
\value{
a \code{data.frame}
}
\description{
Produce DNA concentration data in a clean format for use in R.
}
\examples{
BioDIGS_DNA_conc_data()
}
