% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{BioDIGS_metadata}
\alias{BioDIGS_metadata}
\title{Produce data in a nice clean format for use in R.}
\usage{
BioDIGS_metadata()
}
\value{
a \code{data.frame}
}
\description{
Produce data in a nice clean format for use in R.
}
\examples{
BioDIGS_metadata()
}
