% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{BioDIGS_note}
\alias{BioDIGS_note}
\title{BioDIGS CLI alert}
\usage{
BioDIGS_note(x = NULL, website = F)
}
\value{
A command-line alert
}
\description{
Tell users to see metadata, go to the BioDIGS website, etc.
}
\examples{
BioDIGS_note("?BioDIGS_soil_data()")
BioDIGS_note(website = T)
}
\keyword{internal}
