% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{BioDIGS_soil_data}
\alias{BioDIGS_soil_data}
\title{Produce soil testing data in a clean format for use in R.}
\usage{
BioDIGS_soil_data()
}
\value{
a \code{data.frame}
}
\description{
Produce soil testing data in a clean format for use in R.
}
\examples{
BioDIGS_soil_data()
}
