% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{BioDIGS_soil_warning}
\alias{BioDIGS_soil_warning}
\title{BioDIGS Soil CLI warning}
\usage{
BioDIGS_soil_warning()
}
\value{
A command-line alert
}
\description{
Tell users about important caveat in the measurement of arsenic and cadmium.
}
\examples{
BioDIGS_soil_warning()
}
\keyword{internal}
