% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Gets relevant data from Google Sheets.}
\usage{
getdata(
  auth_method = gs4_auth(scopes = "https://www.googleapis.com/auth/spreadsheets")
)
}
\value{
an uncleaned \code{data.frame}
}
\description{
Gets relevant data from Google Sheets.
}
\examples{
getdata()
}
\keyword{internal}
