% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{BioDIGS_metadata}
\alias{BioDIGS_metadata}
\title{Produce data in a nice clean format for use in R.}
\usage{
BioDIGS_metadata(info = TRUE)
}
\arguments{
\item{info}{if set to TRUE, print important information to the console. This can be disabled with \code{info = FALSE}.}
}
\value{
a \code{data.frame}
}
\description{
Produce data in a nice clean format for use in R.
}
\details{
\tabular{lll}{
   \strong{Field Name} \tab  \tab \strong{Description} \cr
   site_id \tab \verb{  } \tab Unique letter and number site name \cr
   type \tab \verb{  } \tab Management type. Can be managed or unmanaged \cr
   latitude \tab \verb{  } \tab Latitude of the sampling location \cr
   longitude \tab \verb{  } \tab Longitude of the sampling location \cr
}
}
\examples{
BioDIGS_metadata()
}
