% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{BioDIGS_metadata}
\alias{BioDIGS_metadata}
\title{Produce data in a nice clean format for use in R.}
\usage{
BioDIGS_metadata(info = TRUE)
}
\arguments{
\item{info}{if set to TRUE, print important information to the console. This can be disabled with \code{info = FALSE}.}
}
\value{
a \code{data.frame}
}
\description{
Produce data in a nice clean format for use in R.
}
\examples{
BioDIGS_metadata()
}
